refactor(mesh): separate constants
