+++ packages/mesh/app/_components/nav.tsx
@@ -6,6 +6,7 @@ import { usePathname } from 'next/navigation'
 
 import Icon, { IconName } from '@/components/icon'
 import InteractiveIcon from '@/components/interactive-icon'
+import { mobileNavIcons, nonMobileNavIcons } from '@/constants/layout'
 
 type IconInfo = {
   icon: {
@@ -16,78 +17,6 @@ type IconInfo = {
   href: string
   text: string
 }
-
-const NonMobileNavIcons = {
-  first: [
-    {
-      icon: {
-        default: 'icon-popular-lg',
-        hover: 'icon-popular-lg-hover',
-        on: 'icon-popular-lg-on',
-      },
-      href: '/popular',
-      text: '熱門',
-    },
-    {
-      icon: {
-        default: 'icon-social-lg',
-        hover: 'icon-social-lg-hover',
-        on: 'icon-social-lg-on',
-      },
-      href: '/social',
-      text: '社群',
-    },
-    {
-      icon: {
-        default: 'icon-latest-lg',
-        hover: 'icon-latest-lg-hover',
-        on: 'icon-latest-lg-on',
-      },
-      href: '/latest',
-      text: '最新',
-    },
-  ],
-  second: [
-    {
-      icon: {
-        default: 'icon-profile-lg',
-        hover: 'icon-profile-lg-hover',
-        on: 'icon-profile-lg-on',
-      },
-      href: '/profile',
-      text: '個人檔案',
-    },
-    {
-      icon: {
-        default: 'icon-wallet-lg',
-        hover: 'icon-wallet-lg-hover',
-        on: 'icon-wallet-lg-on',
-      },
-      href: '/wallet',
-      text: '錢包',
-    },
-    {
-      icon: {
-        default: 'icon-bookmark-lg',
-        hover: 'icon-bookmark-lg-hover',
-        on: 'icon-bookmark-lg-on',
-      },
-      href: '/bookmark',
-      text: '書籤',
-    },
-  ],
-  third: [
-    {
-      icon: {
-        default: 'icon-setting-lg',
-        hover: 'icon-setting-lg-hover',
-        on: 'icon-setting-lg-on',
-      },
-      href: '/setting',
-      text: '設定',
-    },
-  ],
-} as const
 
 const NonMobileNavIcon = ({
   isOn,
@@ -151,7 +80,7 @@ const NonMobileNav = ({
         <div className="py-10">
           {/* top first section */}
           <div className="flex flex-col border-b sm:gap-8 sm:pb-8 md:gap-2 md:pb-5">
-            {NonMobileNavIcons.first.map((iconInfo) => (
+            {nonMobileNavIcons.first.map((iconInfo) => (
               <NonMobileNavIcon
                 key={iconInfo.text}
                 isOn={path === iconInfo.href}
@@ -160,7 +89,7 @@ const NonMobileNav = ({
             ))}
           </div>
           <div className="flex flex-col sm:gap-8 sm:pt-8 md:gap-2 md:pt-5">
-            {NonMobileNavIcons.second.map((iconInfo) => {
+            {nonMobileNavIcons.second.map((iconInfo) => {
               if (iconInfo.text === '個人檔案' && avatarUrl) {
                 return (
                   <NonMobileNavIcon
@@ -184,7 +113,7 @@ const NonMobileNav = ({
         </div>
         {/* bottom (third) part */}
         <div className="flex flex-col border-t py-6">
-          {NonMobileNavIcons.third.map((iconInfo) => (
+          {nonMobileNavIcons.third.map((iconInfo) => (
             <NonMobileNavIcon
               key={iconInfo.text}
               isOn={path === iconInfo.href}
@@ -196,54 +125,6 @@ const NonMobileNav = ({
     </nav>
   )
 }
-
-const MobileNavIcons = [
-  {
-    icon: {
-      default: 'icon-popular',
-      hover: 'icon-popular-hover',
-      on: 'icon-popular-on',
-    },
-    href: '/popular',
-    text: '熱門',
-  },
-  {
-    icon: {
-      default: 'icon-social',
-      hover: 'icon-social-hover',
-      on: 'icon-social-on',
-    },
-    href: '/social',
-    text: '社群',
-  },
-  {
-    icon: {
-      default: 'icon-latest',
-      hover: 'icon-latest-hover',
-      on: 'icon-latest-on',
-    },
-    href: '/latest',
-    text: '最新',
-  },
-  {
-    icon: {
-      default: 'icon-wallet',
-      hover: 'icon-wallet-hover',
-      on: 'icon-wallet-on',
-    },
-    href: '/wallet',
-    text: '錢包',
-  },
-  {
-    icon: {
-      default: 'icon-profile',
-      hover: 'icon-profile-hover',
-      on: 'icon-profile-on',
-    },
-    href: '/profile',
-    text: '個人檔案',
-  },
-] as const
 
 const MobileNavIcon = ({
   isOn,
@@ -300,7 +181,7 @@ const MobileNav = ({
   return (
     <nav className="fixed bottom-0 left-0 right-0 h-[theme(height.nav.default)] border-t bg-white sm:hidden">
       <div className="flex h-full items-center">
-        {MobileNavIcons.map((iconInfo) => (
+        {mobileNavIcons.map((iconInfo) => (
           <MobileNavIcon
             key={iconInfo.icon.default}
             isOn={path === iconInfo.href}
